requirements.txt
altgraph==0.17.4
blinker==1.9.0
bottle==0.13.2
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
clr_loader==0.2.7.post0
colorama==0.4.6
docker==7.1.0
Flask==3.1.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
packaging==24.2
pefile==2023.2.7
pillow==11.1.0
proxy_tools==0.1.0
pycparser==2.22
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2024.11
pythonnet==3.0.5
pywebview==5.3.2
pywin32==308
pywin32-ctypes==0.2.3
requests==2.32.3
setuptools==75.6.0
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3